+++ pkg/analysis_server/test/services/completion/dart/declaration/test_all.dart
@@ -8,6 +8,7 @@ import 'class_member_test.dart' as class_member;
 import 'class_test.dart' as class_;
 import 'closure_test.dart' as closure;
 import 'constructor_test.dart' as constructor_;
+import 'declaring_parameter_test.dart' as declaring_parameter;
 import 'documentation_test.dart' as documentation;
 import 'enum_test.dart' as enum_;
 import 'extension_member_test.dart' as extension_member;
@@ -32,6 +33,7 @@ void main() {
     class_.main();
     closure.main();
     constructor_.main();
+    declaring_parameter.main();
     documentation.main();
     enum_.main();
     extension_member.main();