+++ pkg/analyzer/lib/dart/analysis/analysis_context_collection.dart
@@ -24,13 +24,13 @@ abstract class AnalysisContextCollection {
   /// [dispose] must be invoked after collection is finished being used.
   factory AnalysisContextCollection({
     required List<String> includedPaths,
+    @Deprecated('Use analysis_options.yaml to exclude paths.')
     List<String>? excludedPaths,
     ResourceProvider? resourceProvider,
     String? sdkPath,
   }) {
     return AnalysisContextCollectionImpl(
       includedPaths: includedPaths,
-      excludedPaths: excludedPaths,
       resourceProvider: resourceProvider,
       sdkPath: sdkPath,
     );