+++ pkg/analyzer/test/src/pubspec/diagnostics/asset_path_not_string_test.dart
@@ -2,46 +2,45 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:analyzer/src/diagnostic/diagnostic.dart' as diag;
 import 'package:test_reflective_loader/test_reflective_loader.dart';
 
+import '../../dart/resolution/node_text_expectations.dart';
 import '../pubspec_test_support.dart';
 
 main() {
   defineReflectiveSuite(() {
     defineReflectiveTests(AssetPathNotStringTest);
+    defineReflectiveTests(UpdateNodeTextExpectations);
   });
 }
 
 @reflectiveTest
 class AssetPathNotStringTest extends PubspecDiagnosticTest {
   test_pathIsList() {
-    assertErrors(
-      '''
+    assertDiagnostics('''
 name: sample
 flutter:
   assets:
     - path: [one, two, three]
-''',
-      [diag.assetPathNotString],
-    );
+//          ^^^^^^^^^^^^^^^^^
+// [diag.assetPathNotString] Asset paths are required to be file paths (strings).
+''');
   }
 
   test_pathIsNull() {
-    assertErrors(
-      '''
+    assertDiagnostics('''
 name: sample
 flutter:
   assets:
     - path:
-''',
-      [diag.assetNotString],
-    );
+//        ^
+// [diag.assetNotString][column 11][length 0] Assets are required to be file paths (strings).
+''');
   }
 
   test_pathIsString() {
     newFile('/sample/assets/my_icon.png', '');
-    assertNoErrors('''
+    assertDiagnostics('''
 name: sample
 flutter:
   assets: