+++ pkg/analysis_server_plugin/CHANGELOG.md
@@ -1,7 +1,7 @@
-## 0.3.18-dev
+## 0.3.18
 
-- Require version `13.3.0-dev` of the `analyzer` package.
-- Require version `0.14.12-dev` of the `analyzer_plugin` package.
+- Require version `13.3.0` of the `analyzer` package.
+- Require version `0.14.12` of the `analyzer_plugin` package.
 
 ## 0.3.17