+++ pkg/analyzer_plugin/CHANGELOG.md
@@ -1,6 +1,6 @@
-## 0.14.12-dev
+## 0.14.12
 
-- Require version `13.3.0-dev` of the `analyzer` package.
+- Require version `13.3.0` of the `analyzer` package.
 
 ## 0.14.11